% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SangerAlignment.R
\name{makestring}
\alias{makestring}
\title{makestring function}
\usage{
makestring(x)
}
\description{
For imported sanger files, makestring extracts the sequence + turns into a string.
For gRNA sequences imported as vectors, makestring turns the sequence into a gRNA string.
}
